+++ modules/web/src/de/diedavids/cuba/metadataextensions/web/EntityDialogsImpl.java
@@ -7,8 +7,8 @@ import com.haulmont.cuba.gui.app.core.inputdialog.InputParameter;
 import com.haulmont.cuba.gui.components.ValidationErrors;
 import com.haulmont.cuba.gui.screen.FrameOwner;
 import com.haulmont.cuba.web.AppUI;
-import de.diedavids.cuba.metadataextensions.MetaPropertyInputParameter;
-import de.diedavids.cuba.metadataextensions.MetadataDialogs;
+import de.diedavids.cuba.metadataextensions.EntityAttributeInputParameter;
+import de.diedavids.cuba.metadataextensions.EntityDialogs;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -22,12 +22,12 @@ import java.util.stream.Stream;
 import static com.haulmont.cuba.gui.app.core.inputdialog.InputDialog.INPUT_DIALOG_OK_ACTION;
 
 
-@Component(MetadataDialogs.NAME)
-public class MetadataDialogsImpl implements MetadataDialogs {
+@Component(EntityDialogs.NAME)
+public class EntityDialogsImpl implements EntityDialogs {
 
     @Override
-    public <E extends Entity> MetadataInputDialogBuilder<E> createMetadataInputDialog(FrameOwner frameOwner, Class<E> entityClass) {
-        return new MetadataInputDialogBuilderImpl<E>(frameOwner, entityClass);
+    public <E extends Entity> EntityInputDialogBuilder<E> createEntityInputDialog(FrameOwner frameOwner, Class<E> entityClass) {
+        return new EntityInputDialogBuilderImpl<E>(frameOwner, entityClass);
     }
 
 
@@ -36,65 +36,65 @@ public class MetadataDialogsImpl implements MetadataDialogs {
     }
 
 
-    public class MetadataInputDialogBuilderImpl<E extends Entity> implements MetadataInputDialogBuilder<E> {
+    public class EntityInputDialogBuilderImpl<E extends Entity> implements EntityInputDialogBuilder<E> {
 
         private E entityInstance;
         private String caption;
         private Consumer<InputDialog.InputDialogCloseEvent> listener;
-        private List<MetaPropertyInputParameter> parameters = new ArrayList<>(2);
+        private List<EntityAttributeInputParameter> parameters = new ArrayList<>(2);
 
         private final Dialogs.InputDialogBuilder inputDialog;
         private final Class<? extends Entity> entityClass;
 
-        public <E extends Entity> MetadataInputDialogBuilderImpl(FrameOwner frameOwner, Class<E> entityClass) {
+        public <E extends Entity> EntityInputDialogBuilderImpl(FrameOwner frameOwner, Class<E> entityClass) {
             inputDialog = getDialogs().createInputDialog(frameOwner);
             this.entityClass = entityClass;
         }
 
         @Override
-        public MetadataInputDialogBuilder withParameter(MetaPropertyInputParameter inputParameter) {
+        public EntityInputDialogBuilder withParameter(EntityAttributeInputParameter inputParameter) {
             parameters.add(inputParameter);
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withParameters(MetaPropertyInputParameter... inputParameters) {
+        public EntityInputDialogBuilder withParameters(EntityAttributeInputParameter... inputParameters) {
             parameters.addAll(Arrays.asList(inputParameters));
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withEntity(E entity) {
+        public EntityInputDialogBuilder withEntity(E entity) {
             this.entityInstance = entity;
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withCaption(String caption) {
+        public EntityInputDialogBuilder withCaption(String caption) {
             this.caption = caption;
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withCloseListener(Consumer<InputDialog.InputDialogCloseEvent> listener) {
+        public EntityInputDialogBuilder withCloseListener(Consumer<InputDialog.InputDialogCloseEvent> listener) {
             this.listener = listener;
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withValidator(Function<InputDialog.ValidationContext, ValidationErrors> validator) {
+        public EntityInputDialogBuilder withValidator(Function<InputDialog.ValidationContext, ValidationErrors> validator) {
             inputDialog.withValidator(validator);
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withWidth(String width) {
+        public EntityInputDialogBuilder withWidth(String width) {
             inputDialog.withWidth(width);
             return this;
         }
 
         @Override
-        public MetadataInputDialogBuilder withHeight(String height) {
+        public EntityInputDialogBuilder withHeight(String height) {
             inputDialog.withHeight(height);
             return this;
         }
@@ -117,22 +117,22 @@ public class MetadataDialogsImpl implements MetadataDialogs {
         private InputParameter[] convertParameters() {
             return parameters
                     .stream()
-                    .map(MetaPropertyInputParameter::getInputParameter)
+                    .map(EntityAttributeInputParameter::getInputParameter)
                     .toArray(InputParameter[]::new);
         }
 
         private void assignDefaultValuesToAutoBindingParameters() {
             parametersWithAutoBinding()
                     .forEach(
-                            metaPropertyInputParameter -> metaPropertyInputParameter.withDefaultValue(
-                                    entityInstance.getValue(metaPropertyInputParameter.getMetaProperty().getName())
+                            entityAttributeInputParameter -> entityAttributeInputParameter.withDefaultValue(
+                                    entityInstance.getValue(entityAttributeInputParameter.getMetaProperty().getName())
                             )
                     );
         }
 
-        private Stream<MetaPropertyInputParameter> parametersWithAutoBinding() {
+        private Stream<EntityAttributeInputParameter> parametersWithAutoBinding() {
             return parameters.stream()
-                    .filter(MetaPropertyInputParameter::isAutoBinding);
+                    .filter(EntityAttributeInputParameter::isAutoBinding);
         }
 
         private Consumer<InputDialog.InputDialogCloseEvent> closeListener() {
@@ -153,7 +153,7 @@ public class MetadataDialogsImpl implements MetadataDialogs {
 
         private void updateEntityInstanceValues(Map<String, Object> values) {
             parametersWithAutoBinding()
-                    .map(metaPropertyInputParameter -> metaPropertyInputParameter.getMetaProperty().getName())
+                    .map(entityAttributeInputParameter -> entityAttributeInputParameter.getMetaProperty().getName())
                     .forEach(parameterName ->
                             entityInstance.setValue(parameterName, values.get(parameterName))
                     );